pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-tour"
version = "0.1.8"
description = "MCP server providing Korea Tourism Organization APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.4.1",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.0",
    "certifi>=2024.2.2"
]
authors = [{ name = "tjwodud04", email = "tjwodud04@gmail.com" }]

[project.urls]
Repository = "https://github.com/tjwodud04/mcp-tour.git"

[project.scripts]
mcp-tour-claude = "mcp_tour.hosts.claude_desktop:install_to_claude_desktop"
mcp-tour-cursor = "mcp_tour.hosts.cursor:install_to_cursor"

[tool.hatch.build.targets.wheel]
packages = ["mcp_tour"]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "ruff>=0.11.2",
    "twine>=6.1.0",
]